% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_logo_reactable}
\alias{add_logo_reactable}
\title{Add logo to reactable}
\usage{
add_logo_reactable(table, opts_theme)
}
\arguments{
\item{table}{Reactable object}

\item{path}{Path to logo png/jpg file}

\item{width}{Width of logo in px; default is 150}

\item{height}{Height of logo in px; default is NULL, resizes automatically to width.}
}
\value{
Reactable object with logo appended to htmlwidget
}
\description{
Add logo to reactable
}
