% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{import_google_font}
\alias{import_google_font}
\title{Import google font to raw HTML of the htmlwidget}
\usage{
import_google_font(viz, opts_theme)
}
\arguments{
\item{viz}{Visualization to which font should be added.}

\item{font}{Google font to import.}
}
\value{
htmlwidget with added line of font import
}
\description{
Import google font to raw HTML of the htmlwidget
}
