% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_logo_reactable}
\alias{add_logo_reactable}
\title{Add logo to reactable}
\usage{
add_logo_reactable(table, opts_theme)
}
\arguments{
\item{table}{Reactable object}

\item{opts_theme}{Theme options from dsvizopts}
}
\value{
Reactable object with logo appended to htmlwidget
}
\description{
Add logo to reactable
}
